Only import graphics if Vu+ Multiboot

+++ lib/python/Tools/Multiboot.py
@@ -4,17 +4,19 @@ import shutil
 import subprocess
 import tempfile
 from os import mkdir, path, rmdir, rename, remove, sep, stat, system
-if not path.exists("/usr/bin/ffmpeg"):
-    system('opkg update')
-    system('opkg install ffmpeg')
-from PIL import Image
-from PIL import ImageDraw
-from PIL import ImageFont
 
 from boxbranding import getMachineBuild, getMachineMtdRoot
 from Components.Console import Console
 from Components.SystemInfo import SystemInfo, BoxInfo as BoxInfoRunningInstance, BoxInformation
 from Tools.Directories import fileHas, fileExists
+
+if fileHas("/proc/cmdline", "kexec=1") and not path.exists("/usr/bin/ffmpeg"):
+		system('opkg update')
+		system('opkg install ffmpeg')
+if fileHas("/proc/cmdline", "kexec=1"):		
+	from PIL import Image
+	from PIL import ImageDraw
+	from PIL import ImageFont
 
 MbootList1 = ("/dev/mmcblk0p1", "/dev/mmcblk1p1", "/dev/mmcblk0p3", "/dev/mmcblk0p4", "/dev/mtdblock2", "/dev/block/by-name/bootoptions")
 MbootList2 = ("/dev/%s" % getMachineMtdRoot(), )	# kexec kernel Vu+ multiboot
